Refactor tries counting to use ulong and improve game generation
Changed the type of the 'Tries' property and related variables from int to ulong for better handling of large counts. Updated GameGenerator and GameGeneratorDynamic to use ulong for tries, improved the game generation loop to handle failures and retries, and enhanced file cleanup logic in Program.cs to remove incomplete results on failure.

+++ LaneAssignments.Functionality/Game.cs
@@ -12,9 +12,9 @@ namespace LaneAssignments
 
         public List<Lane> LaneAssignments { get; set; }
 
-        public int Tries { get; set; }
+        public ulong Tries { get; set; }
 
-        public Game(int number, int tries, params Lane[] lanes)
+        public Game(int number, ulong tries, params Lane[] lanes)
         {
             Number = number;
             Tries = tries;
